+++ src/mod.ts
@@ -24,16 +24,13 @@ export type {
   ScanOptionIndexedStartKey,
   ScanOptions,
 } from './scan-options.js';
+export type {HTTPRequestInfo, InitInput, Mutation} from './repm-invoker.js';
+export type {LogLevel} from './logger.js';
 export type {
-  HTTPRequestInfo,
-  InitInput,
   PatchOperation,
   Puller,
   PullerArg,
   PullerResult,
   PullResponse,
-  Pusher,
-  PusherArg,
-  Mutation,
-} from './repm-invoker.js';
-export type {LogLevel} from './logger.js';
+} from './puller.js';
+export type {Pusher, PusherArg} from './pusher.js';
