+++ src/repm-invoker.ts
@@ -2,6 +2,8 @@ import type {JSONValue} from './json.js';
 import type {ScanOptionsRPC} from './scan-options.js';
 import init, {dispatch} from './wasm/release/replicache_client.js';
 import type {InitOutput} from './wasm/release/replicache_client.js';
+import type {Puller} from './puller.js';
+import type {Pusher} from './pusher.js';
 
 /**
  * This type is used for the [[ReplicacheOptions.wasmModule]] property.
@@ -151,40 +153,6 @@ type BeginTryPullRequest = {
   puller: Puller;
 };
 
-export type PullerArg = {
-  clientID: string;
-  cookie: JSONValue;
-  lastMutationID: number;
-  pullVersion: number;
-  schemaVersion: string;
-  url: string;
-  auth: string;
-  requestID: string;
-};
-
-export type PatchOperation =
-  | {
-      op: 'put';
-
-      key: string;
-      value: JSONValue;
-    }
-  | {op: 'del'; key: string}
-  | {op: 'clear'};
-
-export type PullResponse = {
-  cookie: JSONValue;
-  lastMutationID: number;
-  patch: PatchOperation[];
-};
-
-export type PullerResult = {
-  response?: PullResponse;
-  httpRequestInfo: HTTPRequestInfo;
-};
-
-export type Puller = (arg: PullerArg) => Promise<PullerResult>;
-
 type BeginTryPullResponse = {
   httpRequestInfo: HTTPRequestInfo;
   syncHead: string;
@@ -196,18 +164,6 @@ type TryPushRequest = {
   pushAuth: string;
   schemaVersion: string;
   pusher: Pusher;
-};
-
-export type Pusher = (arg: PusherArg) => Promise<HTTPRequestInfo>;
-
-export type PusherArg = {
-  clientID: string;
-  mutations: Mutation[];
-  pushVersion: number;
-  schemaVersion: string;
-  url: string;
-  auth: string;
-  requestID: string;
 };
 
 type TryPushResponse = {
